requirements.txt
Django==5.1.3
python-decouple==3.8
python-dotenv==1.2.2
gunicorn